requirements.txt
pandas
matplotlib
numpy
tqdm
torch
torchvision
argparse
pyyaml